Separate async/await feature and stability tests
